# Command structure rework