--- a/src/commands/anime/kemonomimi.js
+++ b/src/commands/anime/kemonomimi.js
@@ -1,5 +1,4 @@
-const { CommandStructures, SwitchbladeEmbed } = require('../../')
-const { Command } = CommandStructures
+const { Command, SwitchbladeEmbed } = require('../../')
 const snekfetch = require('snekfetch')
 const nekoAPI = 'https://nekos.life/api/v2/img/'