--- a/src/commands/anime/kitsune.js
+++ b/src/commands/anime/kitsune.js
@@ -5,10 +5,11 @@ const nekoAPI = 'https://nekos.life/api/v2/img/'
 
 module.exports = class Kitsune extends Command {
   constructor (client) {
-    super(client)
-    this.name = 'kitsune'
-    this.aliases = ['foxgirl']
-    this.category = 'anime'
+    super(client, {
+      name: 'kitsune',
+      aliases: ['foxgirl'],
+      category: 'anime'
+    })
   }
 
   async run ({ t, author, channel }) {