--- a/src/commands/anime/myanimelist.js
+++ b/src/commands/anime/myanimelist.js
@@ -1,5 +1,4 @@
-const { CommandStructures, SwitchbladeEmbed, Constants } = require('../../')
-const { Command, CommandError, CommandParameters, StringParameter } = CommandStructures
+const { Command, CommandError, SwitchbladeEmbed, Constants } = require('../../')
 const malScraper = require('mal-scraper')
 
 module.exports = class MyAnimeList extends Command {