--- a/src/commands/anime/nekogif.js
+++ b/src/commands/anime/nekogif.js
@@ -5,9 +5,10 @@ const nekoAPI = 'https://nekos.life/api/v2/img/'
 
 module.exports = class NekoGif extends Command {
   constructor (client) {
-    super(client)
-    this.name = 'nekogif'
-    this.category = 'anime'
+    super(client, {
+      name: 'nekogif',
+      category: 'anime'
+    })
   }
 
   async run ({ t, author, channel }) {