--- a/src/commands/bot/dbl.js
+++ b/src/commands/bot/dbl.js
@@ -1,5 +1,4 @@
-const { SwitchbladeEmbed, Constants, MiscUtils } = require('../../')
-const SearchCommand = require('../../structures/command/SearchCommand.js')
+const { SearchCommand, SwitchbladeEmbed, Constants, MiscUtils } = require('../../')
 
 const moment = require('moment')