--- a/src/commands/developers/blacklist.js
+++ b/src/commands/developers/blacklist.js
@@ -3,16 +3,17 @@ const { Command, CommandRequirements, CommandParameters, StringParameter, UserPa
 
 module.exports = class BlacklistCommand extends Command {
   constructor (client) {
-    super(client)
-    this.name = 'blacklist'
-    this.category = 'developers'
-    this.hidden = true
-
-    this.requirements = new CommandRequirements(this, { devOnly: true })
-    this.parameters = new CommandParameters(this,
-      new UserParameter({ acceptDeveloper: false, missingError: 'commands:blacklist.missingUser' }),
-      new StringParameter({ full: true, missingError: 'commands:blacklist.missingReason' })
-    )
+    super(client, {
+      name: 'blacklist',
+      category: 'developers',
+      hidden: true,
+      requirements: { devOnly: true },
+      parameters: [{
+        type: 'user', acceptDeveloper: false, missingError: 'commands:blacklist.missingUser'
+      }, {
+        type: 'string', full: true, missingError: 'commands:blacklist.missingReason'
+      }]
+    })
   }
 
   async run ({ channel, author, t }, user, reason) {