--- a/src/commands/developers/reloadlocales.js
+++ b/src/commands/developers/reloadlocales.js
@@ -2,12 +2,12 @@ const { CommandStructures, SwitchbladeEmbed, Constants } = require('../../')
 const { Command, CommandRequirements } = CommandStructures
 module.exports = class reloadlocales extends Command {
   constructor (client) {
-    super(client)
-    this.name = 'reloadlocales'
-    this.category = 'developers'
-    this.hidden = true
-
-    this.requirements = new CommandRequirements(this, { managersOnly: true })
+    super(client, {
+      name: 'reloadlocales',
+      category: 'developers',
+      hidden: true,
+      requirements: { managersOnly: true }
+    })
   }
 
   async run ({ t, channel, author }) {