--- a/src/commands/developers/unblacklist.js
+++ b/src/commands/developers/unblacklist.js
@@ -3,15 +3,15 @@ const { Command, CommandRequirements, CommandParameters, UserParameter } = Comma
 
 module.exports = class Unblacklist extends Command {
   constructor (client) {
-    super(client)
-    this.name = 'unblacklist'
-    this.category = 'developers'
-    this.hidden = true
-
-    this.requirements = new CommandRequirements(this, { devOnly: true })
-    this.parameters = new CommandParameters(this,
-      new UserParameter({ showUsage: false, missingError: 'commands:unblacklist.missingUser' })
-    )
+    super(client, {
+      name: 'unblacklist',
+      category: 'developers',
+      hidden: true,
+      requirements: { devOnly: true },
+      parameters: [{
+        type: 'user', showUsage: false, missingError: 'commands:unblacklist.missingUser'
+      }]
+    })
   }
 
   async run ({ channel, author, t }, user) {