--- a/src/commands/developers/welcometranslator.js
+++ b/src/commands/developers/welcometranslator.js
@@ -2,14 +2,16 @@ const { CommandStructures } = require('../../')
 const { Command, CommandParameters, CommandRequirements, UserParameter } = CommandStructures
 module.exports = class welcometranslator extends Command {
   constructor (client) {
-    super(client)
-    this.name = 'welcometranslator'
-    this.category = 'developers'
-    this.hidden = true
-
-    this.requirements = new CommandRequirements(this, { managersOnly: true })
-    this.parameters = new CommandParameters(this,
-      new UserParameter({ missingError: 'commands:welcometranslator.noMember', acceptSelf: false, errors: { acceptSelf: 'commands:welcometranslator.cantWelcomeYourself' } }))
+    super(client, {
+      name: 'welcometranslator',
+      category: 'developers',
+      hidden: true,
+      requirements: { managersOnly: true },
+      parameters: [{
+        type: 'user', missingError: 'commands:welcometranslator.noMember',
+        acceptSelf: false, errors: { acceptSelf: 'commands:welcometranslator.cantWelcomeYourself' }
+      }]
+    })
   }
 
   run ({ t, guild, member: author, channel }, member) {