--- a/src/commands/developers/whyblacklisted.js
+++ b/src/commands/developers/whyblacklisted.js
@@ -3,15 +3,15 @@ const { Command, CommandRequirements, CommandParameters, UserParameter } = Comma
 
 module.exports = class WhyBlacklisted extends Command {
   constructor (client) {
-    super(client)
-    this.name = 'whyblacklisted'
-    this.category = 'developers'
-    this.hidden = true
-
-    this.requirements = new CommandRequirements(this, { devOnly: true })
-    this.parameters = new CommandParameters(this,
-      new UserParameter({ showUsage: false, missingError: 'commands:whyblacklisted.missingUser' })
-    )
+    super(client, {
+      name: 'whyblacklisted',
+      category: 'developers',
+      hidden: true,
+      requirements: { devOnly: true },
+      parameters: [{
+        type: 'user', showUsage: false, missingError: 'commands:whyblacklisted.missingUser'
+      }]
+    })
   }
 
   async run ({ channel, author, t }, user) {