--- a/src/commands/economy/daily.js
+++ b/src/commands/economy/daily.js
@@ -1,5 +1,4 @@
-const { CommandStructures, SwitchbladeEmbed, Constants } = require('../../')
-const { Command, CommandRequirements } = CommandStructures
+const { Command, SwitchbladeEmbed, Constants } = require('../../')
 
 module.exports = class Daily extends Command {
   constructor (client) {