--- a/src/commands/games/beatsaver.js
+++ b/src/commands/games/beatsaver.js
@@ -6,14 +6,14 @@ const cheerio = require('cheerio')
 
 module.exports = class BeatSaver extends Command {
   constructor (client) {
-    super(client)
-    this.name = 'beatsaver'
-    this.aliases = ['beatsaber', 'bsaver']
-    this.category = 'games'
-
-    this.parameters = new CommandParameters(this,
-      new StringParameter({ missingError: 'commands:beatsaver.noQuery' })
-    )
+    super(client, {
+      name: 'beatsaver',
+      aliases: ['beatsaber', 'bsaver'],
+      category: 'games',
+      parameters: [{
+        type: 'string', missingError: 'commands:beatsaver.noQuery'
+      }]
+    })
   }
 
   async run ({ t, author, channel }, query) {