--- a/src/commands/games/beatsaver.js
+++ b/src/commands/games/beatsaver.js
@@ -1,5 +1,4 @@
-const { CommandStructures, SwitchbladeEmbed, Constants } = require('../../')
-const { Command, CommandParameters, StringParameter } = CommandStructures
+const { Command, SwitchbladeEmbed, Constants } = require('../../')
 
 const snekfetch = require('snekfetch')
 const cheerio = require('cheerio')