--- a/src/commands/games/betflip.js
+++ b/src/commands/games/betflip.js
@@ -3,13 +3,16 @@ const { Command, CommandParameters, NumberParameter, StringParameter } = Command
 
 module.exports = class Betflip extends Command {
   constructor (client) {
-    super(client)
-    this.name = 'betflip'
-    this.category = 'games'
-    this.parameters = new CommandParameters(this,
-      new NumberParameter({ min: 1, missingError: 'commands:betflip.noBetValue' }),
-      new StringParameter({ full: true, whitelist: ['heads', 'tails'], missingError: 'commands:betflip.noCoin' })
-    )
+    super(client, {
+      name: 'betflip',
+      aliases: ['bf'],
+      category: 'games',
+      parameters: [{
+        type: 'number', min: 1, missingError: 'commands:betflip.noBetValue'
+      }, {
+        type: 'string', full: true, whitelist: ['heads', 'tails'], missingError: 'commands:betflip.noCoin'
+      }]
+    })
   }
 
   async run ({ channel, author, t }, bet, side) {