--- a/src/commands/games/betflip.js
+++ b/src/commands/games/betflip.js
@@ -1,5 +1,4 @@
-const { CommandStructures, SwitchbladeEmbed, Constants } = require('../../')
-const { Command, CommandParameters, NumberParameter, StringParameter } = CommandStructures
+const { Command, SwitchbladeEmbed, Constants } = require('../../')
 
 module.exports = class Betflip extends Command {
   constructor (client) {