… src/commands/games/coinflip.js
… src/commands/games/fivem.js
… src/commands/games/lolstatus.js
… src/commands/games/mcquery.js
… src/commands/games/samp.js
… src/commands/games/steamladder.js
… src/commands/games/steamladder/profile.js
… src/commands/games/steamladder/steamladder.js
… src/commands/games/tictactoe.js
… src/commands/gifs/handshake.js
… src/commands/gifs/hug.js
… src/commands/gifs/kiss.js
… src/commands/gifs/pat.js
… src/commands/gifs/slap.js
… src/commands/image-manipulation/presidentialalert.js
… src/commands/image-manipulation/triggered.js
… src/commands/memes/bolinadegorfe.js
… src/commands/memes/clapify.js
… src/commands/memes/owo.js
… src/commands/memes/reversetext.js
… src/commands/memes/smart.js
… src/commands/memes/tipsfedora.js
… src/commands/misc/8ball.js
… src/commands/misc/adorable.js
… src/commands/misc/asciify.js
… src/commands/misc/cat.js
… src/commands/misc/cow.js
… src/commands/misc/cowsay.js
… src/commands/misc/deezer.js
… src/commands/misc/deezer/album.js