## Proposition
I want to refactor the requirements/parameters system, but before that I feel I should lighten up the command structure. Right now we have to import and instantiate a lot of classes and that's **probably** 
one of the reasons behind the high RAM usage.

**Not so fun fact:** Right now, we have **160** commands and we instantiate **55** `CommandRequirements` classes, **98** `CommandParameters` classes and **107** implementations of `Parameter` class. This gives us an average of **2.625** classes instances per command (including the command itself).

## Planned
- [x] `CommandRequirements`, `CommandParameters` and `Parameter` are now static classes instead of instantiated ones
- [x] Command complexity reducement
  - [x] Reduces command extensions (`SearchCommand`, `RandomRedditPostCommand`) import paths
  - [x] Reduces imports per command
- [x] RAM usage benchmarking
- [x] `MemberParameter` now extends `UserParameter`
- [x] Closes #593 

## Examples
<details><summary>Before</summary>
<p>

```js
class Ban extends Command {
  constructor (client) {
    super(client)
    this.name = 'ban'
    this.aliases = ['banir']
    this.category = 'moderation'

    this.requirements = new CommandRequirements(this, { guildOnly: true, botPermissions: ['BAN_MEMBERS'], permissions: ['BAN_MEMBERS'] })
    this.parameters = new CommandParameters(this,
      new MemberParameter({ acceptBot: true, missingError: 'commands:ban.missingUser' }),
      new StringParameter({ full: true, missingError: 'commands:ban.missingReason' })
    )
  }
}
```

</p>
</details>

<details><summary>After</summary>
<p>

```js
class Ban extends Command {
  constructor (client) {
    super(client, {
      name: 'ban',
      aliases: ['banir'],
      category: 'moderation',
      requirements: { guildOnly: true, botPermissions: ['BAN_MEMBERS'], permissions: ['BAN_MEMBERS'] },
      parameters: [
        { type: 'member', acceptBot: true, missingError: 'commands:ban.missingUser' },
        { type: 'string', full: true, missingError: 'commands:ban.missingReason' }
      ]
    })
  }
}
```

</p>
</details>

## Memory usage benchmarking

<details><summary>1 minute uptime</summary>
<p>

|     | RSS | Heap total | Heap used |
| :--- | :--- | :--- | :--- |
| Before | 114.89 | 78.033 | 68.028 |
| After | 111.137 | 77.509 | 68.390 |

</p>
</details>

<details><summary>5 minutes uptime</summary>
<p>

|     | RSS | Heap total | Heap used |
| :--- | :--- | :--- | :--- |
| Before | 112.124 | 78.557 | 68.651 |
| After | 111.010 | 77.509 | 68.678 |

</p>
</details>

<details><summary>10 minutes uptime</summary>
<p>

|     | RSS | Heap total | Heap used |
| :--- | :--- | :--- | :--- |
| Before | 111.903 | 78.557 | 68.823 |
| After | 111.342 | 78.033 | 69.004 |

</p>
</details>