@pedrofracassi 2019-02-21
Added checkboxes to the **Planned** section.